Add reports routes
