+++ lib/routes/pastes.js
@@ -80,33 +80,10 @@ router.post('/:id/pastaPoints', async (request, response) => {
   }
 });
 
-router.post('/api/report/:pasteId', async (request, response) => {
+router.post('/:id/decrypt', async (request, response) => {
   try {
-    const pasteId = request.params.pasteId;
-    const paste = await getPaste(pasteId);
-    const reportIcon = '🚨REPORT🚨';
-    const pasteLink = `${request.headers.host}/pastes/${paste._id}`;
-    const slackMessage = {
-      text: `${reportIcon} \n===\n${paste.author}: \n"${paste.value}"\n=== \nID: ${paste._id} \n${pasteLink} \n${reportIcon}`
-    };
-
-    await fetch(process.env.SLACK_WEBHOOK_URL, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(slackMessage)
-    });
-    response.end('Notification sent');
-  } catch (error) {
-    console.error(error);
-    response.status(400).end('Error');
-  }
-});
-
-router.post('/api/decrypt', async (request, response) => {
-  try {
-    const { pasteId, password } = request.body;
+    const pasteId = request.params.id;
+    const { password } = request.body;
 
     const paste = await getPaste(pasteId);
     const decryptedValue = decrypt(paste.value, password);
